refactor: centralize name utils
